Update hts221_reg.c
Hello,
I tried to use this driver with the X-NUCLEO-IKS01A1 board to read the HTS221 sensor values. In my case the register values of HTS221_H0_T0_OUT_L and HTS221_T0_OUT_L are 0xFFFF which means -1 (int16). But as float value I got 65535. In my opinion this should be completely wrong value. To make my project working I added the temporary variable to functions to get the right int16 value and after that for further calculations I converted it to the float value.
Best regards Stanislav